Simplify effects and organise files and simplify types!
