+++ src/settingInternal.ts
