+++ src/types.ts
@@ -3,27 +3,19 @@ import { RepondTypes } from "./declarations";
 // https://stackoverflow.com/questions/49401866/all-possible-keys-of-an-union-type
 export type KeysOfUnion<T> = T extends any ? keyof T : never;
 
-export type InnerEffectCheck<
-  T_State extends { [key: string]: any },
-  T_ItemType extends string
-> =
+export type InnerEffectCheck<T_State extends { [key: string]: any }, T_ItemType extends string> =
   | {
       types?: [T_ItemType];
       names?: string[];
-      props?: KeysOfUnion<
-        NonNullable<T_State[T_ItemType]>[keyof T_State[T_ItemType]]
-      >[];
+      props?: KeysOfUnion<NonNullable<T_State[T_ItemType]>[keyof T_State[T_ItemType]]>[];
       addedOrRemoved?: boolean;
     }
   | {
       types?: T_ItemType;
       names?: string[];
-      props?: KeysOfUnion<
-        NonNullable<T_State[T_ItemType]>[keyof T_State[T_ItemType]]
-      >[];
+      props?: KeysOfUnion<NonNullable<T_State[T_ItemType]>[keyof T_State[T_ItemType]]>[];
       addedOrRemoved?: boolean;
     };
-//
 
 // https://stackoverflow.com/a/55930310 Readonly object , Ben Carp
 
@@ -41,14 +33,7 @@ export type DeepReadonly<T> =
     ? DeepReadonlyObject<T>
     : T;
 
-export type Primitive =
-  | null
-  | undefined
-  | string
-  | number
-  | boolean
-  | symbol
-  | bigint;
+export type Primitive = null | undefined | string | number | boolean | symbol | bigint;
 
 export type AnyFunction = (...args: any[]) => any;
 
@@ -58,8 +43,7 @@ type DeepReadonlyObject<T> = {
 
 interface IDRMap<K, V> extends ReadonlyMap<DeepReadonly<K>, DeepReadonly<V>> {}
 
-interface ReadonlySetDeep<ItemType>
-  extends ReadonlySet<DeepReadonly<ItemType>> {}
+interface ReadonlySetDeep<ItemType> extends ReadonlySet<DeepReadonly<ItemType>> {}
 //
 
 // ----------------------------
@@ -67,9 +51,7 @@ interface ReadonlySetDeep<ItemType>
 
 // type Without<T, K> = Pick<T, Exclude<keyof T, K>>;
 type WithoutB<T, U> = { [P in Exclude<keyof T, keyof U>]?: never };
-export type XOR<T, U> = T | U extends object
-  ? (WithoutB<T, U> & U) | (WithoutB<U, T> & T)
-  : T | U;
+export type XOR<T, U> = T | U extends object ? (WithoutB<T, U> & U) | (WithoutB<U, T> & T) : T | U;
 // NOTE: could use ts-xor package (same)
 
 export type RecursivePartial<T> = { [P in keyof T]?: RecursivePartial<T[P]> };
@@ -101,9 +83,7 @@ export type ExtendsString<T> = T extends string ? T : never;
 export type RepondCallback = (frameDuration: number, frameTime: number) => any;
 
 export type SetRepondState<T_State> = (
-  newState:
-    | GetPartialState<T_State>
-    | ((state: DeepReadonly<T_State>) => GetPartialState<T_State> | undefined),
+  newState: GetPartialState<T_State> | ((state: DeepReadonly<T_State>) => GetPartialState<T_State> | undefined),
   callback?: RepondCallback
 ) => void;
 
@@ -115,31 +95,25 @@ export type StepName = RepondTypes["StepNames"][number];
 type AllStoreInfo = RepondTypes["AllStoreInfo"]; // TODO rename?
 export type ItemType = keyof AllStoreInfo;
 
-export type DefaultStates = {
-  [K_Type in ItemType]: AllStoreInfo[K_Type]["state"];
-};
-export type DefaultRefs = {
-  [K_Type in ItemType]: AllStoreInfo[K_Type]["refs"];
-};
-type Get_DefaultRefs<K_Type extends keyof AllStoreInfo> =
-  AllStoreInfo[K_Type]["refs"];
+export type DefaultStates = { [K_Type in ItemType]: AllStoreInfo[K_Type]["state"] };
+export type DefaultRefs = { [K_Type in ItemType]: AllStoreInfo[K_Type]["refs"] };
+type Get_DefaultRefs<K_Type extends keyof AllStoreInfo> = AllStoreInfo[K_Type]["refs"];
 
 // Make a type that has the start states of all the stores
 export type StartStates = {
   [K_Type in ItemType]: AllStoreInfo[K_Type]["startStates"];
 };
 
-export type StartStatesItemName<K_Type extends keyof AllStoreInfo> =
-  AllStoreInfo[K_Type]["startStates"] extends Record<string, any>
-    ? keyof AllStoreInfo[K_Type]["startStates"]
-    : string;
+export type StartStatesItemName<K_Type extends keyof AllStoreInfo> = AllStoreInfo[K_Type]["startStates"] extends Record<
+  string,
+  any
+>
+  ? keyof AllStoreInfo[K_Type]["startStates"]
+  : string;
 
 // make an AllState type that conditionally uses the keys and values of startStates if available or otherwise uses string as the key and the return type of the default "state"  (for that store) as the value
 export type AllState = {
-  [K_Type in ItemType]: AllStoreInfo[K_Type]["startStates"] extends Record<
-    string,
-    any
-  >
+  [K_Type in ItemType]: AllStoreInfo[K_Type]["startStates"] extends Record<string, any>
     ? AllStoreInfo[K_Type]["startStates"]
     : Record<string, ReturnType<AllStoreInfo[K_Type]["state"]>>;
 };
@@ -152,33 +126,15 @@ export type AllRefs = {
   >;
 };
 
-export type ItemName<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = ExtendsString<KeysOfUnion<T_State[K_Type]>>;
-
-export type PropName<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = KeysOfUnion<T_State[K_Type][ItemName<K_Type, T_ItemType, T_State>]> &
-  string;
-
-export type AllProps<
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
-  [K_Type in T_ItemType]: PropName<K_Type, T_ItemType, T_State>;
-}[T_ItemType];
+export type ItemName<K_Type extends ItemType> = ExtendsString<KeysOfUnion<AllState[K_Type]>>;
+export type PropName<K_Type extends ItemType> = KeysOfUnion<AllState[K_Type][ItemName<K_Type>]> & string;
+export type AllProps = { [K_Type in ItemType]: PropName<K_Type> }[ItemType];
 
 type OriginalGroupNames = keyof RepondTypes["GroupedEffects"];
 type RefinedGroupNames = RemoveEffectsSuffix<OriginalGroupNames>;
 
 // Helper type to strip "Effects" suffix from group names
-type RemoveEffectsSuffix<T extends string> = T extends `${infer Prefix}Effects`
-  ? Prefix
-  : T;
+type RemoveEffectsSuffix<T extends string> = T extends `${infer Prefix}Effects` ? Prefix : T;
 
 export type RefinedGroupedEffects = {
   [K in keyof RepondTypes["GroupedEffects"] as RemoveEffectsSuffix<K>]: RepondTypes["GroupedEffects"][K];
@@ -187,82 +143,49 @@ export type RefinedGroupedEffects = {
 // ------------------------------------------------------------
 // DiffInfo
 
-type DiffInfo_PropsChanged<
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
-  [K_Type in T_ItemType]: Record<
-    ItemName<K_Type, T_ItemType, T_State>,
-    PropName<K_Type, T_ItemType, T_State>[]
-  > & {
-    all__: PropName<K_Type, T_ItemType, T_State>[];
-  };
-} & {
-  all__: AllProps<T_ItemType, T_State>[];
-};
+type DiffInfo_PropsChanged = {
+  [K_Type in ItemType]: Record<ItemName<K_Type>, PropName<K_Type>[]> & { all__: PropName<K_Type>[] };
+} & { all__: AllProps[] };
 
-type DiffInfo_PropsChangedBool<
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
-  [K_Type in T_ItemType]: Record<
-    ItemName<K_Type, T_ItemType, T_State>,
-    { [K_PropName in PropName<K_Type, T_ItemType, T_State>]: boolean }
-  > & {
+type DiffInfo_PropsChangedBool = {
+  [K_Type in ItemType]: Record<ItemName<K_Type>, { [K_PropName in PropName<K_Type>]: boolean }> & {
     all__: {
-      [K_PropName in PropName<K_Type, T_ItemType, T_State>]: boolean;
+      [K_PropName in PropName<K_Type>]: boolean;
     };
   };
 } & {
-  all__: { [K_PropName in AllProps<T_ItemType, T_State>]: boolean };
+  all__: { [K_PropName in AllProps]: boolean };
 };
 
-type DiffInfo_ItemsChanged<
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = Record<T_ItemType | "all__", ItemName<T_ItemType, T_ItemType, T_State>[]>;
+type DiffInfo_ItemsChanged = Record<ItemType | "all__", ItemName<ItemType>[]>;
 
-type DiffInfo_ItemsChangedBool<
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = Record<
-  T_ItemType | "all__",
-  Record<ItemName<T_ItemType, T_ItemType, T_State>, boolean>
->;
+type DiffInfo_ItemsChangedBool = Record<ItemType | "all__", Record<ItemName<ItemType>, boolean>>;
 
-export type DiffInfo<
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
-  itemTypesChanged: T_ItemType[];
-  itemsChanged: DiffInfo_ItemsChanged<T_ItemType, T_State>;
-  propsChanged: DiffInfo_PropsChanged<T_ItemType, T_State>;
-  itemsAdded: DiffInfo_ItemsChanged<T_ItemType, T_State>;
-  itemsRemoved: DiffInfo_ItemsChanged<T_ItemType, T_State>;
-  itemTypesChangedBool: Record<T_ItemType | "all__", boolean>;
-  itemsChangedBool: DiffInfo_ItemsChangedBool<T_ItemType, T_State>;
-  propsChangedBool: DiffInfo_PropsChangedBool<T_ItemType, T_State>;
-  itemsAddedBool: DiffInfo_ItemsChangedBool<T_ItemType, T_State>;
-  itemsRemovedBool: DiffInfo_ItemsChangedBool<T_ItemType, T_State>;
+export type DiffInfo = {
+  itemTypesChanged: ItemType[];
+  itemsChanged: DiffInfo_ItemsChanged;
+  propsChanged: DiffInfo_PropsChanged;
+  itemsAdded: DiffInfo_ItemsChanged;
+  itemsRemoved: DiffInfo_ItemsChanged;
+  itemTypesChangedBool: Record<ItemType | "all__", boolean>;
+  itemsChangedBool: DiffInfo_ItemsChangedBool;
+  propsChangedBool: DiffInfo_PropsChangedBool;
+  itemsAddedBool: DiffInfo_ItemsChangedBool;
+  itemsRemovedBool: DiffInfo_ItemsChangedBool;
 };
 
 // ------------------------------------------------------------
 
-export type ItemEffect_Run_Params<
-  K_Type extends T_ItemType,
-  K_PropName extends PropName<K_Type, T_ItemType, T_State>,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>,
-  T_Refs extends Record<any, any>
-> = {
-  itemName: ItemName<K_Type, T_ItemType, T_State>;
-  newValue: T_State[K_Type][ItemName<K_Type, T_ItemType, T_State>][K_PropName];
-  prevValue: T_State[K_Type][ItemName<K_Type, T_ItemType, T_State>][K_PropName];
-  itemState: T_State[K_Type][ItemName<K_Type, T_ItemType, T_State>];
+export type ItemEffect_Run_Params<K_Type extends ItemType, K_PropName extends PropName<K_Type>> = {
+  itemName: ItemName<K_Type>;
+  newValue: AllState[K_Type][ItemName<K_Type>][K_PropName];
+  prevValue: AllState[K_Type][ItemName<K_Type>][K_PropName];
+  itemState: AllState[K_Type][ItemName<K_Type>];
   // itemRefs: T_Refs[K_Type][keyof T_Refs[K_Type]];
-  itemRefs: T_Refs[K_Type][ItemName<K_Type, T_ItemType, T_State>];
+  itemRefs: AllRefs[K_Type][ItemName<K_Type>];
   // itemRefs: Get_T_Refs<K_Type>[ItemName<K_Type>];
   frameDuration: number;
+  ranWithoutChange?: boolean;
 };
 
 export type ItemEffect_Check_Becomes =
@@ -274,148 +197,82 @@ export type ItemEffect_Check_Becomes =
 
 // for useStoreItem
 
-type OneItem_OneCheck_OneProperty<
-  K_Type extends T_ItemType,
-  K_PropName extends PropName<K_Type, T_ItemType, T_State>,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
+type OneItem_OneCheck_OneProperty<K_Type extends ItemType, K_PropName extends PropName<K_Type>> = {
   prop?: K_PropName;
   type: K_Type;
-  name: ItemName<K_Type, T_ItemType, T_State>;
+  name: ItemName<K_Type>;
   becomes?: ItemEffect_Check_Becomes;
   addedOrRemoved?: undefined;
 };
 
-type OneItem_ACheck_MultiProperties<
-  K_Type extends T_ItemType,
-  K_PropName extends PropName<K_Type, T_ItemType, T_State>,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
+type OneItem_ACheck_MultiProperties<K_Type extends ItemType, K_PropName extends PropName<K_Type>> = {
   prop?: K_PropName[];
   type: K_Type; // maybe ideally optional (and handle adding effect with any item type)
-  name: ItemName<K_Type, T_ItemType, T_State>;
+  name: ItemName<K_Type>;
   becomes?: ItemEffect_Check_Becomes;
   addedOrRemoved?: undefined;
 };
 
-export type OneItem_Check<
-  K_Type extends T_ItemType,
-  K_PropName extends PropName<K_Type, T_ItemType, T_State>,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> =
-  | OneItem_OneCheck_OneProperty<K_Type, K_PropName, T_ItemType, T_State>
-  | OneItem_ACheck_MultiProperties<K_Type, K_PropName, T_ItemType, T_State>;
+export type OneItem_Check<K_Type extends ItemType, K_PropName extends PropName<K_Type>> =
+  | OneItem_OneCheck_OneProperty<K_Type, K_PropName>
+  | OneItem_ACheck_MultiProperties<K_Type, K_PropName>;
 
 // -----------------
 //
 
 // for itemEffectCallback
-type ItemEffect_Check_SingleProperty<
-  K_Type extends T_ItemType,
-  K_PropName extends PropName<K_Type, T_ItemType, T_State>,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
+type ItemEffect_Check_SingleProperty<K_Type extends ItemType, K_PropName extends PropName<K_Type>> = {
   prop?: K_PropName;
   type: K_Type;
-  name?:
-    | ItemName<K_Type, T_ItemType, T_State>[]
-    | ItemName<K_Type, T_ItemType, T_State>;
+  name?: ItemName<K_Type>[] | ItemName<K_Type>;
   becomes?: ItemEffect_Check_Becomes;
   addedOrRemoved?: undefined;
 };
 
-type ItemEffect_Check_MultiProperties<
-  K_Type extends T_ItemType,
-  K_PropName extends PropName<K_Type, T_ItemType, T_State>,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
+type ItemEffect_Check_MultiProperties<K_Type extends ItemType, K_PropName extends PropName<K_Type>> = {
   prop?: K_PropName[];
   type: K_Type; // maybe ideally optional (and handle adding listener with any item type)
-  name?:
-    | ItemName<K_Type, T_ItemType, T_State>[]
-    | ItemName<K_Type, T_ItemType, T_State>;
+  name?: ItemName<K_Type>[] | ItemName<K_Type>;
   becomes?: ItemEffect_Check_Becomes;
   addedOrRemoved?: undefined;
 };
 
-export type ItemEffect_Check<
-  K_Type extends T_ItemType,
-  K_PropName extends PropName<K_Type, T_ItemType, T_State>,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> =
-  | ItemEffect_Check_SingleProperty<K_Type, K_PropName, T_ItemType, T_State>
-  | ItemEffect_Check_MultiProperties<K_Type, K_PropName, T_ItemType, T_State>;
+export type ItemEffect_Check<K_Type extends ItemType, K_PropName extends PropName<K_Type>> =
+  | ItemEffect_Check_SingleProperty<K_Type, K_PropName>
+  | ItemEffect_Check_MultiProperties<K_Type, K_PropName>;
 
-export type ItemEffect_Run<
-  K_Type extends T_ItemType,
-  K_PropName extends PropName<K_Type, T_ItemType, T_State>,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>,
-  T_Refs extends Record<any, any>
-> = (
-  loopedInfo: ItemEffect_Run_Params<
-    K_Type,
-    K_PropName,
-    T_ItemType,
-    T_State,
-    T_Refs
-  >
+export type ItemEffect_Run<K_Type extends ItemType, K_PropName extends PropName<K_Type>> = (
+  loopedInfo: ItemEffect_Run_Params<K_Type, K_PropName>
 ) => void;
 
-export type ItemEffect_Options<
-  K_Type extends T_ItemType,
-  K_PropName extends PropName<K_Type, T_ItemType, T_State>,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>,
-  T_Refs extends Record<any, any>,
-  T_StepName extends string
-> = {
-  check: ItemEffect_Check<K_Type, K_PropName, T_ItemType, T_State>;
+export type ItemEffect_Options<K_Type extends ItemType, K_PropName extends PropName<K_Type>> = {
+  check: ItemEffect_Check<K_Type, K_PropName>;
   // can use function to check value, ideally it uses the type of the selected property
-  run: ItemEffect_Run<K_Type, K_PropName, T_ItemType, T_State, T_Refs>;
+  run: ItemEffect_Run<K_Type, K_PropName>;
   atStepEnd?: boolean;
   name?: string;
-  step?: T_StepName;
+  step?: StepName;
   runAtStart?: boolean;
   _isPerItem?: true;
 };
 
-export type ItemEffect_Options_NoMeta<
-  K_Type extends T_ItemType,
-  K_PropName extends PropName<K_Type, T_ItemType, T_State>,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>,
-  T_Refs extends Record<any, any>,
-  T_StepName extends string
-> = {
-  check: ItemEffect_Check<K_Type, K_PropName, T_ItemType, T_State>;
+export type ItemEffect_Options_NoMeta<K_Type extends ItemType, K_PropName extends PropName<K_Type>> = {
+  check: ItemEffect_Check<K_Type, K_PropName>;
   // can use function to check value, ideally it uses the type of the selected property
-  run: ItemEffect_Run<K_Type, K_PropName, T_ItemType, T_State, T_Refs>;
+  run: ItemEffect_Run<K_Type, K_PropName>;
   atStepEnd?: boolean;
   name?: string;
-  step?: T_StepName;
+  step?: StepName;
   runAtStart?: boolean;
 };
 
 // -----------------
 // AnyChangeRule ( like a slightly different and typed repond listener )
 
-type Effect_OneCheck_OneItemType<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
+type Effect_OneCheck_OneItemType<K_Type extends ItemType> = {
   type?: K_Type;
-  name?:
-    | ItemName<K_Type, T_ItemType, T_State>
-    | ItemName<K_Type, T_ItemType, T_State>[];
-  prop?: PropName<K_Type, T_ItemType, T_State>[];
+  name?: ItemName<K_Type> | ItemName<K_Type>[];
+  prop?: PropName<K_Type>[];
   addedOrRemoved?: boolean;
   becomes?: undefined;
 };
@@ -429,136 +286,68 @@ type Effect_OneCheck_OneItemType<
   // "characters" and "global" need to be in array array so
   // the MultiItemTypes type's chosen
   */
-type Effect_OneCheck_MultiItemTypes<
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
-  type?: T_ItemType[];
-  name?:
-    | ItemName<T_ItemType, T_ItemType, T_State>
-    | ItemName<T_ItemType, T_ItemType, T_State>[];
-  prop?: AllProps<T_ItemType, T_State>[];
+type Effect_OneCheck_MultiItemTypes = {
+  type?: ItemType[];
+  name?: ItemName<ItemType> | ItemName<ItemType>[];
+  prop?: AllProps[];
   addedOrRemoved?: boolean;
   becomes?: undefined;
 };
 
-export type Effect_OneCheck<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> =
-  | Effect_OneCheck_OneItemType<K_Type, T_ItemType, T_State>
-  | Effect_OneCheck_MultiItemTypes<T_ItemType, T_State>;
+export type Effect_OneCheck<K_Type extends ItemType> =
+  | Effect_OneCheck_OneItemType<K_Type>
+  | Effect_OneCheck_MultiItemTypes;
 
-export type Effect_Check<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> =
-  | Effect_OneCheck<K_Type, T_ItemType, T_State>[]
-  | Effect_OneCheck<K_Type, T_ItemType, T_State>;
+export type Effect_Check<K_Type extends ItemType> = Effect_OneCheck<K_Type>[] | Effect_OneCheck<K_Type>;
 
-export type Effect_Run<
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = (
-  diffInfo: DiffInfo<T_ItemType, T_State>,
+export type Effect_Run = (
+  diffInfo: DiffInfo,
   frameDuration: number,
-  skipChangeCheck?: boolean // for useStoreItemEffect, to run the effect regardless of changes
+  ranWithoutChange?: boolean // for useStoreItemEffect, to run the effect regardless of changes
 ) => void;
 
-export type Effect_Options<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>,
-  T_StepName extends string
-> = {
+export type Effect_Options<K_Type extends ItemType> = {
   name?: string; // ruleName NOTE used to be required (probably still for dangerouslyAddingRule ? (adding single rules without making rules first))
-  check: Effect_Check<K_Type, T_ItemType, T_State>;
-  run: Effect_Run<T_ItemType, T_State>;
+  check: Effect_Check<K_Type>;
+  run: Effect_Run;
   atStepEnd?: boolean;
-  step?: T_StepName;
+  step?: StepName;
   runAtStart?: boolean;
   _isPerItem?: false;
 };
 
-export type Effect_Options_NoMeta<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>,
-  T_StepName extends string
-> = {
+export type Effect_Options_NoMeta<K_Type extends ItemType> = {
   name?: string; // effectName NOTE used to be required (probably still for startNewEffect
-  check: Effect_Check<K_Type, T_ItemType, T_State>;
-  run: Effect_Run<T_ItemType, T_State>;
+  check: Effect_Check<K_Type>;
+  run: Effect_Run;
   atStepEnd?: boolean;
-  step?: T_StepName;
+  step?: StepName;
   runAtStart?: boolean;
 };
 
-export type Effect_Options_NameRequired<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>,
-  T_StepName extends string
-> = Effect_Options<K_Type, T_ItemType, T_State, T_StepName> & {
-  name: string;
-};
+export type Effect_Options_NameRequired<K_Type extends ItemType> = Effect_Options<K_Type> & { name: string };
 
-type AnyEffect_Options<
-  K_Type extends T_ItemType,
-  K_PropName extends PropName<K_Type, T_ItemType, T_State>,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>,
-  T_Refs extends Record<any, any>,
-  T_StepName extends string
-> = XOR<
-  Effect_Options<K_Type, T_ItemType, T_State, T_StepName>,
-  ItemEffect_Options<
-    K_Type,
-    K_PropName,
-    T_ItemType,
-    T_State,
-    T_Refs,
-    T_StepName
-  >
+type AnyEffect_Options<K_Type extends ItemType, K_PropName extends PropName<K_Type>> = XOR<
+  Effect_Options<K_Type>,
+  ItemEffect_Options<K_Type, K_PropName>
 >;
 
-export type MakeRule_Rule<
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>,
-  T_Refs extends Record<any, any>,
-  T_StepName extends string
-> = AnyEffect_Options<
-  T_ItemType,
-  PropName<T_ItemType, T_ItemType, T_State>,
-  T_ItemType,
-  T_State,
-  T_Refs,
-  T_StepName
->;
+export type MakeRule_Rule = AnyEffect_Options<ItemType, PropName<ItemType>>;
 
 // ----------------------------
 //  Listener types
 
-type InnerEffect_OneCheck_OneItemType<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
+type InnerEffect_OneCheck_OneItemType<K_Type extends ItemType> = {
   types?: K_Type;
-  names?: ItemName<K_Type, T_ItemType, T_State>[];
-  props?: PropName<K_Type, T_ItemType, T_State>;
+  names?: ItemName<K_Type>[];
+  props?: PropName<K_Type>;
   addedOrRemoved?: boolean;
 };
 
-export type InnerEffect_OneCheck_MultiItemTypes<
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
-  types?: (keyof T_State)[];
-  names?: ItemName<T_ItemType, T_ItemType, T_State>[];
-  props?: AllProps<T_ItemType, T_State>[];
+export type InnerEffect_OneCheck_MultiItemTypes = {
+  types?: (keyof AllState)[];
+  names?: ItemName<ItemType>[];
+  props?: AllProps[];
   addedOrRemoved?: boolean;
 };
 
@@ -578,60 +367,35 @@ export type InnerEffect_OneCheck_MultiItemTypes<
 //   TestChangeToCheckUnionWithoutProperties<T, K>
 // >;
 
-export type InnerEffect_OneCheck<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> =
-  | InnerEffect_OneCheck_OneItemType<K_Type, T_ItemType, T_State>
-  | InnerEffect_OneCheck_MultiItemTypes<T_ItemType, T_State>;
+export type InnerEffect_OneCheck<K_Type extends ItemType> =
+  | InnerEffect_OneCheck_OneItemType<K_Type>
+  | InnerEffect_OneCheck_MultiItemTypes;
 
-export type InnerEffect_Check<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> =
-  | InnerEffect_OneCheck<K_Type, T_ItemType, T_State>[]
-  | InnerEffect_OneCheck<K_Type, T_ItemType, T_State>;
+export type InnerEffect_Check<K_Type extends ItemType> = InnerEffect_OneCheck<K_Type>[] | InnerEffect_OneCheck<K_Type>;
 
-export type AddItem_OptionsUntyped<
-  T_State extends Record<any, any>,
-  T_Refs extends Record<any, any>,
-  T_TypeName
-> = {
+export type AddItem_OptionsUntyped<T_State extends Record<any, any>, T_Refs extends Record<any, any>, T_TypeName> = {
   type: string;
   name: string;
-  state?: Partial<
-    NonNullable<T_State[T_TypeName]>[keyof T_State[keyof T_State]]
-  >;
+  state?: Partial<NonNullable<T_State[T_TypeName]>[keyof T_State[keyof T_State]]>;
   refs?: Partial<NonNullable<T_Refs[T_TypeName]>[keyof T_Refs[keyof T_Refs]]>;
 };
 
 // -----------------
 // Listeners B
 
-export type InnerEffect_Loose<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>,
-  T_StepName extends string
-> = {
+export type InnerEffect_Loose<K_Type extends ItemType> = {
   name: string;
-  check: InnerEffect_Check<K_Type, T_ItemType, T_State>;
-  run: (diffInfo: DiffInfo<T_ItemType, T_State>, frameDuration: number) => void;
+  check: InnerEffect_Check<K_Type>;
+  run: (diffInfo: DiffInfo, frameDuration: number) => void;
   atStepEnd?: boolean;
-  step?: T_StepName;
+  step?: StepName;
 };
 
 // After normalizing
-export type InnerEffect<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>
-> = {
+export type InnerEffect<K_Type extends ItemType> = {
   name: string;
-  check: InnerEffect_OneCheck<K_Type, T_ItemType, T_State>[];
-  run: (diffInfo: DiffInfo<T_ItemType, T_State>, frameDuration: number) => void;
+  check: InnerEffect_OneCheck<K_Type>[];
+  run: (diffInfo: DiffInfo, frameDuration: number) => void;
   atStepEnd?: boolean;
   step?: string;
 };
@@ -639,16 +403,11 @@ export type InnerEffect<
 // -----------------
 //
 
-export type UseStoreItem_Params<
-  K_Type extends T_ItemType,
-  T_ItemType extends string | number | symbol,
-  T_State extends Record<any, any>,
-  T_Refs extends Record<any, any>
-> = {
-  itemName: ItemName<K_Type, T_ItemType, T_State>;
-  prevItemState: T_State[K_Type][ItemName<K_Type, T_ItemType, T_State>];
-  itemState: T_State[K_Type][ItemName<K_Type, T_ItemType, T_State>];
+export type UseStoreItem_Params<K_Type extends ItemType> = {
+  itemName: ItemName<K_Type>;
+  prevItemState: AllState[K_Type][ItemName<K_Type>];
+  itemState: AllState[K_Type][ItemName<K_Type>];
   // itemRefs: T_Refs[K_Type][keyof T_Refs[K_Type]];
-  itemRefs: T_Refs[K_Type][ItemName<K_Type, T_ItemType, T_State>];
+  itemRefs: AllRefs[K_Type][ItemName<K_Type>];
   // frameDuration: number;
 };
